Split GetFeatureInfo methods out of data.py and spit ogc_utils.py,
