+++ datacube_ows/ows_configuration.py
@@ -40,7 +40,8 @@ from datacube_ows.config_utils import (FlagProductBands, OWSConfigEntry,
                                        get_file_loc, import_python_obj,
                                        load_json_obj, ConfigException, FunctionWrapper, CFG_DICT, RAW_CFG)
 from datacube_ows.cube_pool import ODCInitException, cube, get_cube
-from datacube_ows.ogc_utils import create_geobox, local_solar_date_range
+from datacube_ows.ogc_utils import create_geobox
+from datacube_ows.time_utils import local_solar_date_range
 from datacube_ows.resource_limits import OWSResourceManagementRules, parse_cache_age
 from datacube_ows.styles import StyleDef
 from datacube_ows.tile_matrix_sets import TileMatrixSet
