+++ datacube_ows/resource_limits.py
@@ -12,8 +12,8 @@ from odc.geo.geobox import GeoBox
 from odc.geo.geom import CRS, polygon
 
 from datacube_ows.config_utils import CFG_DICT, RAW_CFG, OWSConfigEntry, ConfigException
-from datacube_ows.ogc_utils import (cache_control_headers,
-                                    create_geobox)
+from datacube_ows.ogc_utils import (create_geobox)
+from datacube_ows.http_utils import cache_control_headers
 
 TYPE_CHECKING = False
 if TYPE_CHECKING:
