+++ datacube_ows/wms.py
@@ -5,10 +5,11 @@
 # SPDX-License-Identifier: Apache-2.0
 from flask import render_template
 
-from datacube_ows.data import feature_info, get_map
+from datacube_ows.data import get_map
+from datacube_ows.feature_info import feature_info
 from datacube_ows.legend_generator import legend_graphic
 from datacube_ows.ogc_exceptions import WMSException
-from datacube_ows.ogc_utils import cache_control_headers, get_service_base_url
+from datacube_ows.http_utils import get_service_base_url, cache_control_headers
 from datacube_ows.ows_configuration import get_config
 from datacube_ows.utils import log_call
