+++ tests/test_data.py
@@ -12,7 +12,8 @@ from odc.geo.geom import polygon
 from xarray import Dataset
 
 import datacube_ows.data
-from datacube_ows.data import get_s3_browser_uris
+import datacube_ows.feature_info
+from datacube_ows.feature_info import get_s3_browser_uris
 from datacube_ows.loading import ProductBandQuery, DataStacker
 from datacube_ows.ogc_exceptions import WMSException
 from tests.test_styles import product_layer  # noqa: F401
@@ -135,7 +136,7 @@ def test_make_derived_band_dict_nan():
         "fake": fake_style()
     }
 
-    band_dict = datacube_ows.data._make_derived_band_dict(fake_dataset(), style_dict)
+    band_dict = datacube_ows.feature_info._make_derived_band_dict(fake_dataset(), style_dict)
     assert band_dict["fake"] == "n/a"
 
 
@@ -162,7 +163,7 @@ def test_make_derived_band_dict_not_nan():
         "fake": fake_style()
     }
 
-    band_dict = datacube_ows.data._make_derived_band_dict(fake_dataset(), style_dict)
+    band_dict = datacube_ows.feature_info._make_derived_band_dict(fake_dataset(), style_dict)
     assert band_dict["fake"] == 10.10
 
 
@@ -186,7 +187,7 @@ def test_make_band_dict_nan(product_layer): # noqa: F811
 
     bands = ["fake"]
 
-    band_dict = datacube_ows.data._make_band_dict(product_layer, fake_dataset())
+    band_dict = datacube_ows.feature_info._make_band_dict(product_layer, fake_dataset())
     assert band_dict["fake"] == "n/a"
 
 
@@ -232,13 +233,13 @@ def test_make_band_dict_float(product_layer): # noqa: F811
 
     bands = ["fake"]
 
-    band_dict = datacube_ows.data._make_band_dict(product_layer, int_dataset())
+    band_dict = datacube_ows.feature_info._make_band_dict(product_layer, int_dataset())
     assert isinstance(band_dict["fake"], dict)
     assert band_dict["fake"] == {
         "Mask image as provided by JAXA - Ocean and water, lay over, shadowing, land.": 'lay_over'
     }
 
-    band_dict = datacube_ows.data._make_band_dict(product_layer, float_dataset())
+    band_dict = datacube_ows.feature_info._make_band_dict(product_layer, float_dataset())
     assert isinstance(band_dict["fake"], dict)
     assert band_dict["fake"] == {
         "Mask image as provided by JAXA - Ocean and water, lay over, shadowing, land.": 'lay_over'
